requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckeditor==6.7.1
django-ckeditor-5==0.2.14
django-cors-headers==4.3.0
django-dotenv==1.4.2
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg==3.1.12
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
